+++ src/main/java/com/EstoqueFacil/EstoqueFacil/repository/LoteRepository.java
@@ -10,9 +10,9 @@ import com.EstoqueFacil.EstoqueFacil.model.Lote;
 @Repository
 public interface LoteRepository extends JpaRepository<Lote, Integer> {
 
-    Optional<Lote> findByNumeroLote(int numeroLote);
-    boolean existsByNumeroLote(int numeroLote);
-    void deleteByNumeroLote(int numeroLote);
+    Optional<Lote> findByNumeroLote(long numeroLote);
+    boolean existsByNumeroLote(long numeroLote);
+    void deleteByNumeroLote(long numeroLote);
 
 }
